fix(demo): move clinet component to own file
